requirements.txt
altgraph==0.17
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.6
future==0.18.2
image==1.5.33
imutils==0.5.4
MouseInfo==0.1.3
mss==2.0.22
numpy==1.21.1
opencv-python==4.5.3.56
pefile==2021.5.24
Pillow==8.3.1
psutil==5.8.0
PyAutoGUI==0.9.53
pycodestyle==2.7.0
PyGetWindow==0.0.9
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
PyMsgBox==1.0.9
pynput==1.6.8
pyocr==0.8
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.27
pytesseract==0.3.8
PyTweening==1.0.3
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
